hak akses

+++ app/Http/Controllers/tahapanPerkembanganDataController.php
@@ -11,6 +11,9 @@ class TahapanPerkembanganDataController extends Controller
 {
     public function index()
     {
+        if (auth()->user()->role !== 'orangtua') {
+            abort(403, 'Unauthorized');
+        }
         // Mengambil pencapaian tahapan perkembangan yang dimiliki user yang sedang login
         $tahapanPerkembanganData = TahapanPerkembanganData::where('user_id', Auth::id())->get();
 
@@ -19,6 +22,9 @@ class TahapanPerkembanganDataController extends Controller
 
     public function create()
     {
+        if (auth()->user()->role !== 'orangtua') {
+            abort(403, 'Unauthorized');
+        }
         // Menampilkan daftar tahapan perkembangan untuk dipilih oleh orang tua
         $tahapanPerkembangan = TahapanPerkembangan::all();
         return view('orangtua.tahapan_perkembangan.create', compact('tahapanPerkembangan'));
@@ -26,6 +32,9 @@ class TahapanPerkembanganDataController extends Controller
 
     public function edit($id)
     {
+        if (auth()->user()->role !== 'orangtua') {
+            abort(403, 'Unauthorized');
+        }
         // Mengambil data tahapan perkembangan berdasarkan ID
         $tahapanPerkembanganData = TahapanPerkembanganData::findOrFail($id);
         $tahapanPerkembangan = TahapanPerkembangan::all();
@@ -36,6 +45,9 @@ class TahapanPerkembanganDataController extends Controller
 
     public function store(Request $request)
     {
+        if (auth()->user()->role !== 'orangtua') {
+            abort(403, 'Unauthorized');
+        }
         // Validasi data yang diterima
         $request->validate([
             'tahapan_perkembangan_id' => 'required|exists:tahapan_perkembangan,id',
@@ -58,6 +70,9 @@ class TahapanPerkembanganDataController extends Controller
 
     public function update(Request $request, $id)
     {
+        if (auth()->user()->role !== 'orangtua') {
+            abort(403, 'Unauthorized');
+        }
         // Validasi data yang diterima
         $request->validate([
             'tahapan_perkembangan_id' => 'required|exists:tahapan_perkembangan,id',
@@ -81,6 +96,9 @@ class TahapanPerkembanganDataController extends Controller
     }
     public function destroy($id)
     {
+        if (auth()->user()->role !== 'orangtua') {
+            abort(403, 'Unauthorized');
+        }
         $tahapanPerkembanganData = TahapanPerkembanganData::findOrFail($id);
         $tahapanPerkembanganData->delete();
